refactor(tests): the oracle accessors belong on Scratch, and the helpers were tripled
/simplify pass over this branch. Three findings, all reuse.

changes_of was defined in fidelity.rs and budgets.rs, and this branch made it
three. numbered_lines was in fidelity.rs and this branch made it two. Both move
to support/, which is where the shared fixture helpers already live, and the
sorted spelling is now named changes_sorted so it cannot be confused with the
unsorted one budgets.rs keeps deliberately: that suite measures the walk, so
reordering what the walk reported would measure something else.

git_numstat and checkout were free functions in normalise.rs taking &Scratch,
which is a method with extra steps. Both are now Scratch methods, and
git_numstat sits beside git_hunk_headers because it is the same job: ask git the
question the engine just answered. SPEC.md section 10 names git as the oracle,
so the accessors that reach it live together.

No behaviour change. fidelity.rs shrinks by its two copies, normalise.rs by four
helpers. 344 tests still pass.
